+++ lua/obsidian/commands/init.lua
@@ -7,6 +7,7 @@ local command_lookups = {
   ObsidianToday = "obsidian.commands.today",
   ObsidianYesterday = "obsidian.commands.yesterday",
   ObsidianTomorrow = "obsidian.commands.tomorrow",
+  ObsidianDailies = "obsidian.commands.dailies",
   ObsidianNew = "obsidian.commands.new",
   ObsidianOpen = "obsidian.commands.open",
   ObsidianBacklinks = "obsidian.commands.backlinks",
@@ -147,6 +148,8 @@ M.register("ObsidianToday", { opts = { nargs = "?", desc = "Open today's daily n
 M.register("ObsidianYesterday", { opts = { nargs = 0, desc = "Open the daily note for the previous working day" } })
 
 M.register("ObsidianTomorrow", { opts = { nargs = 0, desc = "Open the daily note for the next working day" } })
+
+M.register("ObsidianDailies", { opts = { nargs = "*", desc = "Open a picker with daily notes" } })
 
 M.register("ObsidianNew", { opts = { nargs = "?", desc = "Create a new note" } })